Support to limit the number of pending outgoing messages
When there's no limitation, the library itself would queue all messages
with QoS > 0 if the messages can not be delivered. This can make the
queue grow to an unreasonable size and lead an out-of-memory issue.
For example, if the user keep publishing messages when the messages
cannot be sent successfully.

Signed-off-by: Chang Yu-heng <mr.changyuheng@gmail.com>