Add extra argument for websocket auth headers
Used mainly for connecting to the AWS IoT service over websockets. This requires extra headers with the hash of the normal headers and the AWS secret access key. This commit allows passing in a function when creating the client object that will then return any extra headers that will be used to authenticate with the server.

Signed-off-by: Michael Boulton <boulton@zoetrope.io>