Make auth dict handling a bit more leniant
This commit changes the logic slightly on handling of the auth dict that
gets passed in to publish or subscribe methods. Previously they expected
None to always be what is passed in which was the default. However if an
empty dict was passed in this would fail with a KeyError. Additionally
the same was true if you pased a dict in without a username key. To make
this a bit more forgiving this changes the logic so if an object is
passed in that evaluates to False (like an empty dict) the object isn't
used. It also makes checking for the presence of 'username' and
'password' use the get() method so we don't raise blank Exceptions and
are a bit more explicit in usage.

This should be fully backwards compatible with any previous usage, but
hopefully make it a bit simpler for people playing with the lib for the
first time.

Signed-off-by: Matthew Treinish <mtreinish@kortar.org>