Leaving write_loop() call to socket_register_write callback if present.
Instead of calling write_loop() directly when using an external event
loop, check whether the socket_register_write callback is set, and use
it if it is available.

This fixes #361, since on_publish() is now called from the
socket_register_write_callback that runs in a different
thread/coroutine instead of by the publish() function itself.

Signed-off-by: Jan Seeger <jan.seeger@thenybble.de>